Update style.scss
Adjusted the width of the titles as well as the perspective transformation so the titles fill the screen better now that the reveal starts from the bottom. Before, the action would be more towards the bottom and therefor feels awkward. The new values should also be responsive since they use responsive units (vw/vh).